Prune down values for OpenLDAP
TL;DR
-----

Only set values we change when installing OpenLDAP

Details
-------

I originally used the full `values.yaml` from the OpenLDAP Helm
chart as a baseline file that I edited for my install. This meant
I set a lot of values to their defaults and wouldn't be able to
take advataged of changes to those defaults (or added values) in
newer version fo the chart.

This change fixes that, but only setting the values I want to
control. This is part of shifting all the values files in the
P
codebase to that model.