Prune down values for Vault (#4)
TL;DR
-----

Only set values we change when installing Vault

Details
-------

I originally used the full `values.yaml` from the Vault Helm
chart as a baseline file that I edited for my install. This meant
I set a lot of values to their defaults and wouldn't be able to
take advataged of changes to those defaults (or added values) in
newer version fo the chart.

This change fixes that, but only setting the values I want to
control. This is part of shifting all the values files in the
codebase to that model.