Refactor supervisor cluster namespaces (#6)
TL;DR
-----

Use multiple namespaces for pods in the supervisor cluster to
facilitate automation

Details
-------

Split the original `operations` namespace into two to simplify
Vault automation. It's difficult to automate Vault setup in the
same run of Terraform that creates the Vault server. Since the
convention for this code is one run of Terraform, the simplest
fix was the split Vault into it's own namespace.

Along the way, I also added some automation for the supervisor
cluster in the `terraform/supervior` directory. Supporting GCP
object are now created there. I attempted to also create a real
certificate for the cluster with that automation, but I was
thwarted because the CSR includes and email and Let's Encrypt
won't allow an email in the request. More to come if I can dig
up something on the vCenter server.

Specific changes:

* Move Vault into it's own namespace called `secrets` and add
  the necessary config to set it up.
* Update the `certificate` Terraform module to take a CSR in
  hopes of using it for the supervisor cluster. Hopes were
  dashed but the enhancement remains.
* Add some Terraform that applies to the entire supervisor
  cluster. Right its fairly lightweight and just sets up the
  GCP service accounts and KMS key ring that the automation
  for the various namespaces needs.
* Update automation for the `operations` namespace to refer
  to the remote state of the supervisor cluster automation
  as needed.