Fix refactored references (and a couple of others) (#12)
TL;DR
------

Follow-on Terraforms had some invalid references after the infrastructure refactoring and this fixed them

Details
-------

When I refactored the former supervisor terraform into the infrastructure Terraform I made a few changes to the other Terraform but apparently not enough of them. The secrets terraform still had references to `supervisor` and also hadn't adjusted references to the certificate module and the secret that was moved out of the module. The operations terraform also needed some corrections for certificate references.

While I was in there, I also updated the terraform so that the infrastructure terraform was the source for the subdomain and any google specific properties rather than each one defining it's own variables and locals based on a common convention. Now it only has to be changed in one places if a change is desired.