Use the correct minio hostname (#16)
TL;DR
-----

Fix issue with Minio showing an invalid certificate

Details
-------

Minio was showing a certificate for `local.lab.crdant.net` 
rather than `s3.lab.crdant.net`. That was the certificate that
Terraform created since the host name was set incorrectly.
This change updates the value to the correct name.