Create the tools cluster with Terraform (#17)
TL;DR
-----

Use Terraform to create a cluster for shared tooling

Details
-------

Uses Terraform to drive the `tkg` CLI to create a cluster for
share tools. Right now, there are two resources of type 
`null_resource` that use a local provisioner to first set up
the `tkg` CLI to use the supervisdor cluster, and then to 
create the cluster with `tkg create cluster`.

The second of these null resources was intended to use a 
`kubernetes_manifest` resource, but as of now the supervisor
cluster is not compatible with that approach. Getting there
requires that either the Tanzu Kubernetes Cluster controller
be updated to support a server-side dry run, or that the 
alpha Terraform provider for Kubernetes manifests be changed
to support a client side plan. 

Since my lab is behind by one patch level and there is an 
open issue with the manifest terrform provider, it's possible
that one or both of these issues will be resolved soon.