Don't use Terraform for the tools cluster (#19)
TL;DR
-----

Remove Terraform that created the tools cluster

Details
-------

Creating the tools cluster with Terrform was a bit cludgey, and it
started to feel like a bit more trouble that it was worth. There
were two main challenges:

1. There wasn't any way to work with TKG other than to use a
   null resources and local provisioner. That's always a bit of
   a code smell but seemed workable...until it seemed less so.
2. Connecting to the tools cluster in the same Terraform that
   created it created circular references that were stretching
   what Terraform is capable of. Eventually it was clear that
   it wasn't going to work.

It was possible to split it out into it's own seperate stage of
Terraforming, but that started to feel awkward when the only
things in it were going to be null resource hacks.